See #28. Moves SPI into new file.
* SPI code in spi.h and spi.cpp
* Makefile got a makeover
* Future TODO: don't pass 'spifd' every time we call SPI fn
